chore: move drop handling to a separate component
